Add two WebSocket endpoints
Add two WebSocket endpoints:
- /v1/ws/backupvolumes
- /v1/ws/backups

Haven't found a way to allow passing the volume name as a parameter to filter
per-backup volume's backups change. Either by:
- `/v1/ws/backups/{volName}`
- `/v1/ws/backups?volName=<volName>`
- `/v1/ws/backupvolumes/{backupName}`

So, the current behavior for the `backups` WebSocket endpoint is that
sending all the Backup CRs to the frontend and frontend filter this
backups volume's backup by volumeName filed.

Longhorn #2740

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>