Track terminating pods
We changed the behavior so that we only set lastPodRefAt once there are
no more pods present, previously this was already set when the last pod
went into terminating state.

I added a unit test for both cases, to verify the behavior of tracking in
terminating state and lastPodRefAt setting once there are no more pods
present for that volume.

Signed-off-by: Joshua Moody <joshua.moody@suse.com>