Remove the keyword register
The register keyword in the value of the upper example will enable the
registration of deepcopy methods into the scheme. This will completely
go away in Kubernetes 1.9 because the scheme won’t be responsible anymore
to do deepcopies of runtime.Objects.
Instead just call yourobject.DeepCopy() or yourobject.DeepCopyObject().
You can and you should do that already today in 1.8-based projects as
it is faster and less error-prone.
Moreover, you will be prepared for 1.9 which will require this pattern.

Ref to https://www.openshift.com/blog/kubernetes-deep-dive-code-generation-customresources

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>