controller: Remove invalid migration replicas
If the old replica becomes invalid (mode ERR), it means the data
may be corrupted. Then the migration engine should not continue
using the corresponding migration replica.

And when the new replica becomes ERR while the old replica is
still state running, the controller should remove the failed
new replica then recreate it as well.

Longhorn 2306

Signed-off-by: Shuo Wu <shuo.wu@suse.com>