Do not print error log when backup in progress
If we have 2 cluster points to the same backup store.
When cluster 1 makes snapshot backup to the remote backup store and
cluster 2 starts to inspect the backup config from the remote backup, it
will print the error message that the backup is still in progress.

We should skip printing the in-progress backup to the console.

Longhorn 3402

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>