example: add longhorn webhook server network policy yaml manifest
The traffic from/to the longhorn webhook server is the kube-apiserver.
The only way we could add restriction is to add the network policy of
the ingress port because we can't know each Kubernetes distro default
kube-apiserver Pod's label. Therefore, we can't add the label selector
in the network policy rule to restrict the traffic that comes from the
kube-apiserver is able to access to the longhorn webhook server.

Longhorn 3513

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>