Fix scalability issue
* Increase the resync for all shared informers so that we don't
overwhelm the controller workqueues. One exception is the node
shared informer, we keep resync to be 30s to check the real time
usage of disk on nodes
* When adding a fresh event to wokqueue, we should use Add() instead
of AddRateLimited(). AddRateLimited() should only be used when a controller
encounters error while processing an event and the controller want to add
the event back to its workqueue for later retry. See example at
https://github.com/kubernetes-csi/external-attacher/blob/master/pkg/controller/controller.go

Longhorn-3083

Signed-off-by: Phan Le <phan.le@suse.com>