Add a field to record the longhorn engine generated backup name
The longhorn manager >= 1.2.x + longhorn engine >= 1.2.0, the backup CR name is generated in
longhorn manager and pass to the longhorn engine.
The longhorn manager >= 1.2.x + longhorn engine < 1.2.0, the backup CR name is generated in
longhorn manager but the longhorn engine would generate another name.
To keep backward compatibility of longhorn manager >= 1.2.x + longhorn engine < 1.2.0, stores
the longhorn engine reports backup name as `engineBackupName`, and use this name to query the
longhorn engine to monitor the snapshot backup process.

Note that, with longhorn manager >= 1.2.x + longhorn engine < 1.2.0, the backup CR would be
replaced to a newer one (with the name generated by the longhorn engine) after the snapshot backup
is done and synced from the remote backup target.

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>