backup: retry deletion after failing to acquire the lock
The resync period of the backup is one hour and the maxRetries is 3.
Thus, the deletion failure of the backup in error state is caused by the
shutdown of the replica during backing up, if the lock hold by the backup
job is not released.

The workaround is to increase the maxRetries number and to retry the deletion
until the lock acquisition.

maxRetriesOnAcquireLockError should guarantee the cumulative retry time
is larger than 150 seconds.

Signed-off-by: Derek Su <derek.su@suse.com>