Always fetch the engine backup name and replica address
The key to query the engine's backup status is the engine backup name +
replica address.
If the backup monitor resource disappear, call the engine's SnapshotBackup
to get the engine backup name and replica address.

If the engine's snapshot backup is in progress, calling the engine's
SnapshotBackup would not create a new backup.

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>