Use exponential backoff retry mechanism to fetch engine backup status
If the backup state is in Pending state, which means the engine's
snapshot backup encounters RPC deadline exceeded error.
In this situation, use an exponential backoff retry timer with maximum
retry period 24 hours. The exponential backoff retry time is
4s/8s/16s/.../10mins. Once the snapshot backup found inside the longhorn
engine/replica, the backup monitor changes to liner timer.

If the backup state is not in Pending state, the backup monitor uses
liner timer to periodically collects the longhorn engine/replica backup
status.

Signed-off-by: JenTing Hsiao <jenting.hsiao@suse.com>