Reducing the workqueue depth while still keeping the resync period as 30s
For each controller, we only need to enable resync for the handler
that handles the events of its managed object. For example, the
volume controller only needs to enable resync for volumeInformer
event handler. By doing this, we reduce the workqueue depth of the
controller while still making sure that we resync every object
of the controller every resync period (30s)

Signed-off-by: Phan Le <phan.le@suse.com>