replica: do not check storage IP when in deletion
Do not check storage IP when the replica is marked
for deletion to unblock the flow.

When the replica controller node is down during a
replica deletion, another replica controller
should pick up and finish the flow. However, before
removing the finalizer, the new replica controller
could get blocked by the ownership change when the
storage IP is not in sync.

Longhorn-4213

Signed-off-by: Chin-Ya Huang <chin-ya.huang@suse.com>