Remove forced node assignment for share-manager pod
Let the kubernetes scheduler take care of pod placement, in the future
we want to add nodeSelector support as well. The reason we don't want to
bypass the scheduler are conditions like a node already has max pods.

Withe forced assignment the share-manager pod would never come up and be
stuck forever in a failure scenario.

Signed-off-by: Joshua Moody <joshua.moody@suse.com>