Add support for using encrypted volumes in block mode directly
I use the staging path as a parent folder and bind mount device mapper
mapped encrypted volume device into a file inside of that folder.

This allows us to store additional metadata and give us more control over
our tear down for example since we are not sharing control with kubernetes
as is the case for the staging path itself.

We should be doing the same for the volumes used in fs mode, instead of
directly mounting into the staging path.

Longhorn 4883

Signed-off-by: Joshua Moody <joshua.moody@suse.com>