Make sure pod.Spec.NodeName is assigned before check the node status
Previously, the immediate IsNodeDownOrDeleted check could fail if the share manager pod's spec.nodeName was not set,
which could lead to the detachment of the volume and the transfer of ownership.
A race condition between the deletion of old volume replicas and the attachment of new volume could lead to a faulty volumes.

Longhorn 5537

Signed-off-by: Derek Su <derek.su@suse.com>