Support encrypted volumes in block mode
Use the staging path as a parent folder and bind mount device mapper mapped
encrypted volume device into a file inside of that folder.

This allows us to store additional metadata and give us more control over
our tear down for example since we are not sharing control with kubernetes
as is the case for the staging path itself.

Ref: https://github.com/longhorn/longhorn-manager/pull/1613

Longhorn 4883

Signed-off-by: Derek Su <derek.su@suse.com>