object store: Implement Delete
Implement a lifecycle state machine for the object endpoint resource.
This machine is used by the controller to track the creation of
resources and application of labels and annotations where necessary to
control the lifecycle of the dependant object correctly

Implement uninstall hooks via the uninstall controller. This ensures
that when Longhorn is uninstalled no object endpoint resources are left
lingering in the cluster.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>