fix(restore): dr volume failed by delete lock
DR volume will fail to start an incremental restoration in target
cluster, when the source cluster is deleting the eliminated backup
by deleting lock in the backup store.

We handle the full restoration at the issue 3055 and now add a
regular expression pattern to handle similar failed message to have
a backoff mechanism to retry the full/incremental restoration.

Ref: 6750

Signed-off-by: James Lu <james.lu@suse.com>