object store: fix deletes, remove ingress
Remove ingres for the object store UI. This is no longer needed when
using the Longhorn UI's nginx as reverse proxy.

Fix deleting an object store by removing the finalizer without waiting
on resources to be deleted. Since OwnerReferences are used for automatic
cleanup, the finalizer can be removed immediately.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>