object store: fix race condition, error handling
Fix the race condition on start of a new object store. As long as there
is just the object store and no longhorn volume yet some controller
needs to be responsible for the object store. Usually it would be the
one responsible for the volume, this doesn't exist yet and there are no
provisions to track the owner another way. To fix this, just let the
controller on the first node handle it.

Fix error handling by setting the object store state to error only once
and otherwise propagate errors up to the reconcile function.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>