object store: reorder checks ; clean up
- Reorder the health checks during the creation of an object store

When creating an object store, the responsibility for managing the
resources making up that object store rests with the controller that's
also responsible for the longhorn volume.
There is a brief moment while the object store is starting when the
information which controller is responsible for the longhorn volume has
not yet been propagated to all controllers. Therefore the controller
must first wait until the longhorn volume is healthy and has an owner
before it can determine for sure if it's responsible or not. Only then
the other resources can be created without running into race conditions
with other controller who think they are responsible.

- Clean up utility functions and constants

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>