object store: rebase and update vendored modules
Rebase on master and update vendored modules

Fix logic in test script. To ensure only the necessary unit tests are
executed (or of nothing has been changed, all unit tests are executed),
the logic needs to make sure not to interprete vendored modules as
changed source since it will fail to find unit tests for them.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>