fix: ignore unknown node condition
Longhorn exposed all unknown (to longhorn) node conditions as K8s
events. This is a mechanism that only has informational value and is
used as such in Longhorn as well, as it only serves to relay
information to the UI.

In particular, the condition in question is `EtcdIsVoter` and is a K3s
specific condition, indicating the membership details of an etcd node on
that K3s node and is irrelevant to Longhorn, but other non-standard node
conditions should be ignored too, if Longhorn doesn't derive any
actionable value from them.

related-to: longhorn/longhorn#7290

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>