Do not create any instance manager pods for v2 data engine if one is already existing
An extra instance manager pod for v2 data engine requires an additional CPU core
and 2 GiB of memory. If the resources are insufficient, the instance manager pod
will be recreated and retried. To minimize unnecessary retries, do not create any
new instance manager pods if one is already running.

Longhorn 8456

Signed-off-by: Derek Su <derek.su@suse.com>