feat(rwx): share manager pod respects the storageClass.allowedTopologies
In the Longhorn storage system, a share manager pod of a rwx volume is created on a random node.
The purpose of this update is to enhance the locality of an RWX volume and its share manager pod.
It accomplishes this by converting the storageClass.allowedTopologies setting into an affinity
configuration for the share manager pod. As a result, both the pod and the RWX volume respect
this affinity, ensuring their allocation to an optimal node.

Longhorn 7872

Signed-off-by: Derek Su <derek.su@suse.com>