crd/instanceManager: add spec.desireState
spec.desireStatus defines the desired state of an instance manager.
For an instance manager that serves v1 data engines, this field is always
set to "running". Conversely, for an instance manager handling v2 data
engines, the field is set to "stopped" if another instance manager is
already running on the same node.

Longhorn 8456

Signed-off-by: Derek Su <derek.su@suse.com>