Fall back to a running instance-manager if a default is not available
Some operations in controllers, such as deleting disks and reconciling
backup targets, initially fetch the default instance-manager. However,
these operations are not restricted to the default instance-manager;
they can use any running instance-manager. To enhance resilience, if
the default instance-manager is unable to start before an upgrade of
the v2 data engine or for any other reason for both v1 and v2 data
engines, the controller can switch to any running instance-manager as
a fallback.

Longhorn 8464

Signed-off-by: Derek Su <derek.su@suse.com>