Update the checkVolumeCreated condition for CSI CreateVolume call
Sometime the volume is auto-attached quickly after being created. For
example, volume which require backup restore or volume require data
cloning. We shouldn't wait for the volume to be detached before return
OK for CreateVolume call because it could take a very long time to
finish the data transfer.

longhorn-2907

Signed-off-by: Phan Le <phan.le@suse.com>