fix: reject snapshot purge when volume is being live migration
Reject snapshot purge when volume is being live migration.
If snapshot purge is executed while live migrating a volume,
the disk files will be changed. The new engine and replicas
are unable to be aware of the change and cannot list the
new disk files, resulting in ERR replicas.

Longhorn 10495

Signed-off-by: Derek Su <derek.su@suse.com>