feat(orphan): introduce orphan-resource-auto-deletion-grace-period setting
In the current implementation, with auto orphan deletion enabled, the orphan CR
and its associated process are deleted almost immediately after the orphan CR is
created. This behavior might be too abrupt. We could introduce a configurable
setting, such as a wait interval (deletion grace period), to delay this automatic
deletion. For instance, if the interval is set to 60 seconds, the orphan CR and
its process would be deleted 60 seconds after the creation timestamp.

Manual deletion of the orphan CR by the user would bypass this delay and proceed
immediately.

Longhorn 10904

Signed-off-by: Derek Su <derek.su@suse.com>