fix: shouldn't count replica marked as deleted as healthy and active replicas
After a replica’s spec.deletionTimestamp is set but before it is
actually deleted, it is still considered a healthy replica.
During the cleanup of extra replicas, the volume controller counts
this replica as healthy and may proceed to delete additional
replicas, potentially resulting in the removal of too many replicas.

Longhorn 11388

Signed-off-by: Derek Su <derek.su@suse.com>